chore(server): move Endpoint to utils
